feat(all): merge path and verify into app.js and conditional render navbar
