+++ client/src/pages/Login.js
@@ -63,7 +63,7 @@ const LandingPage = () => {
       });
       let data = response.data;
       if (data.status === "forgot") {
-        navigate("/verify/otp-verification", { state: email });
+        navigate("/otp-verification", { state: email });
       } else {
         toast.warn("Email not found.");
       }
@@ -80,7 +80,7 @@ const LandingPage = () => {
           <h1>Login</h1>
           <p>
             Don't have an account?{" "}
-            <Link to={"/verify/signup"} className="linksColor">
+            <Link to={"/signup"} className="linksColor">
               {" "}
               Register
             </Link>{" "}
