+++ client/src/pages/Profile.js
@@ -29,7 +29,7 @@ const Profile = () => {
         let data = response.data;
         if (response.status === 401) {
           toast.error("Kindly login first!");
-          navigate("/verify/login");
+          navigate("/login");
           return;
         }
         if (data.status === "ok") {
