+++ client/src/pages/UserProfile.js
@@ -27,7 +27,7 @@ const UserProfile = () => {
       let data = response.data;
       if (response.status === 401) {
         toast.error("Kindly login first!");
-        navigate("/verify/login");
+        navigate("/login");
         return;
       }
       if (data.status === "ok") {
