+++ client/src/pages/Vaccination.js
@@ -61,7 +61,7 @@ const Vaccination = () => {
       let data = response.data;
       if (response.status === 401) {
         toast.error("Kindly login first!");
-        navigate("/verify/login");
+        navigate("/login");
         return;
       }
