+++ client/src/styles/contact.css
@@ -6,7 +6,7 @@
   box-sizing: border-box;
   padding: 0 10%;
   height: 100%;
-  margin-top: 50px;
+  margin-top: 20px;
   min-height: 90vh;
 }
