+++ client/src/styles/guidelines.css
@@ -4,7 +4,7 @@
 }
 
 .communityGuidelinesContainer {
-  margin-top: 100px;
+  margin-top: 85px;
   margin-left: 150px;
   margin-right: 150px;
   font-size: 17px;
